refactor: split implementation by domain
