+++ wire.go
@@ -1,0 +1,300 @@
+package otlpwire
+
+import (
+	"errors"
+	"io"
+	"iter"
+
+	"google.golang.org/protobuf/encoding/protowire"
+)
+
+func repeatedFieldSeq[T ~[]byte](data []byte, fieldNum protowire.Number) (iter.Seq[T], func() error) {
+	var iterErr error
+	seq := func(yield func(T) bool) {
+		forEachRepeatedField(data, fieldNum, func(item []byte, err error) bool {
+			if err != nil {
+				iterErr = err
+				return false
+			}
+			return yield(T(item))
+		})
+	}
+	return seq, func() error { return iterErr }
+}
+
+// skipField skips a field value after its tag has been consumed. The field
+// number is required for groups so ConsumeFieldValue can verify the matching
+// end-group marker and recursively validate its contents.
+func skipField(data []byte, fieldNum protowire.Number, wireType protowire.Type) int {
+	return protowire.ConsumeFieldValue(fieldNum, wireType, data)
+}
+
+// countRepeatedField counts items in a repeated field by delegating to countFunc
+// for each occurrence of the specified field.
+func countRepeatedField(data []byte, fieldNum protowire.Number, countFunc func([]byte) (int, error)) (int, error) {
+	count := 0
+	pos := 0
+
+	for pos < len(data) {
+		num, wireType, tagLen := protowire.ConsumeTag(data[pos:])
+		if tagLen < 0 {
+			return 0, errors.New("malformed protobuf tag")
+		}
+		pos += tagLen
+
+		if num == fieldNum {
+			if wireType != protowire.BytesType {
+				return 0, errors.New("wrong wire type for field")
+			}
+			msgBytes, n := protowire.ConsumeBytes(data[pos:])
+			if n < 0 {
+				return 0, errors.New("invalid bytes in repeated field")
+			}
+			pos += n
+
+			c, err := countFunc(msgBytes)
+			if err != nil {
+				return 0, err
+			}
+			count += c
+		} else {
+			n := skipField(data[pos:], num, wireType)
+			if n < 0 {
+				return 0, errors.New("failed to skip field")
+			}
+			pos += n
+		}
+	}
+
+	return count, nil
+}
+
+// countOccurrences counts direct occurrences of a specific field.
+func countOccurrences(data []byte, fieldNum protowire.Number) (int, error) {
+	count := 0
+	pos := 0
+
+	for pos < len(data) {
+		num, wireType, tagLen := protowire.ConsumeTag(data[pos:])
+		if tagLen < 0 {
+			return 0, errors.New("malformed protobuf tag")
+		}
+		pos += tagLen
+
+		if num == fieldNum {
+			if wireType != protowire.BytesType {
+				return 0, errors.New("wrong wire type for field")
+			}
+			_, n := protowire.ConsumeBytes(data[pos:])
+			if n < 0 {
+				return 0, errors.New("invalid bytes in field")
+			}
+			pos += n
+			count++
+		} else {
+			n := skipField(data[pos:], num, wireType)
+			if n < 0 {
+				return 0, errors.New("failed to skip field")
+			}
+			pos += n
+		}
+	}
+
+	return count, nil
+}
+
+// forEachRepeatedField iterates over a repeated field, calling fn for each occurrence.
+// The callback receives field bytes or an error. Return false to stop iteration.
+func forEachRepeatedField(data []byte, fieldNum protowire.Number, fn func([]byte, error) bool) {
+	pos := 0
+
+	for pos < len(data) {
+		num, wireType, tagLen := protowire.ConsumeTag(data[pos:])
+		if tagLen < 0 {
+			fn(nil, errors.New("malformed protobuf tag"))
+			return
+		}
+		pos += tagLen
+
+		if num == fieldNum {
+			if wireType != protowire.BytesType {
+				fn(nil, errors.New("wrong wire type for field"))
+				return
+			}
+			msgBytes, n := protowire.ConsumeBytes(data[pos:])
+			if n < 0 {
+				fn(nil, errors.New("invalid bytes in repeated field"))
+				return
+			}
+			pos += n
+
+			if !fn(msgBytes, nil) {
+				return
+			}
+		} else {
+			n := skipField(data[pos:], num, wireType)
+			if n < 0 {
+				fn(nil, errors.New("failed to skip field"))
+				return
+			}
+			pos += n
+		}
+	}
+}
+
+// extractResourceMessage extracts the Resource message (field 1) from
+// ResourceMetrics/ResourceLogs/ResourceSpans messages.
+func extractResourceMessage(data []byte) ([]byte, error) {
+	pos := 0
+
+	for pos < len(data) {
+		fieldNum, wireType, tagLen := protowire.ConsumeTag(data[pos:])
+		if tagLen < 0 {
+			return nil, errors.New("malformed protobuf tag")
+		}
+		pos += tagLen
+
+		// Field 1 = Resource (message)
+		if fieldNum == 1 {
+			if wireType != protowire.BytesType {
+				return nil, errors.New("resource field has wrong wire type")
+			}
+			msgBytes, n := protowire.ConsumeBytes(data[pos:])
+			if n < 0 {
+				return nil, errors.New("invalid bytes in resource field")
+			}
+			return msgBytes, nil
+		}
+
+		// Skip other fields
+		n := skipField(data[pos:], fieldNum, wireType)
+		if n < 0 {
+			return nil, errors.New("failed to skip field")
+		}
+		pos += n
+	}
+
+	return nil, errors.New("resource field not found")
+}
+
+// extractBytesField extracts the first occurrence of a length-delimited
+// field from protobuf data. Returns nil (not an error) if absent.
+// The returned slice aliases data; no copy is made.
+func extractBytesField(data []byte, fieldNum protowire.Number) ([]byte, error) {
+	pos := 0
+
+	for pos < len(data) {
+		num, wireType, tagLen := protowire.ConsumeTag(data[pos:])
+		if tagLen < 0 {
+			return nil, errors.New("malformed protobuf tag")
+		}
+		pos += tagLen
+
+		if num == fieldNum {
+			if wireType != protowire.BytesType {
+				return nil, errors.New("wrong wire type for field")
+			}
+			msgBytes, n := protowire.ConsumeBytes(data[pos:])
+			if n < 0 {
+				return nil, errors.New("invalid bytes in field")
+			}
+			return msgBytes, nil
+		}
+
+		n := skipField(data[pos:], num, wireType)
+		if n < 0 {
+			return nil, errors.New("failed to skip field")
+		}
+		pos += n
+	}
+
+	return nil, nil
+}
+
+// extractFixed64Field extracts the first occurrence of a fixed64 field from
+// protobuf data. Returns 0 (not an error) if absent.
+func extractFixed64Field(data []byte, fieldNum protowire.Number) (uint64, error) {
+	pos := 0
+
+	for pos < len(data) {
+		num, wireType, tagLen := protowire.ConsumeTag(data[pos:])
+		if tagLen < 0 {
+			return 0, errors.New("malformed protobuf tag")
+		}
+		pos += tagLen
+
+		if num == fieldNum {
+			if wireType != protowire.Fixed64Type {
+				return 0, errors.New("wrong wire type for field")
+			}
+			v, n := protowire.ConsumeFixed64(data[pos:])
+			if n < 0 {
+				return 0, errors.New("invalid fixed64 in field")
+			}
+			return v, nil
+		}
+
+		n := skipField(data[pos:], num, wireType)
+		if n < 0 {
+			return 0, errors.New("failed to skip field")
+		}
+		pos += n
+	}
+
+	return 0, nil
+}
+
+// writeResourceMessage writes resource data as a valid OTLP export request message.
+// Wraps the resource bytes with field tag 1 and length prefix.
+func writeResourceMessage(w io.Writer, data []byte) (int64, error) {
+	buf := make([]byte, 0, 11) // tag + length varint
+	buf = protowire.AppendTag(buf, 1, protowire.BytesType)
+	buf = protowire.AppendVarint(buf, uint64(len(data)))
+
+	n1, err := w.Write(buf)
+	if err != nil {
+		return int64(n1), err
+	}
+
+	n2, err := w.Write(data)
+	return int64(n1 + n2), err
+}
+
+// extractFixedBytesField extracts a bytes field of known size from protobuf data.
+// Returns nil (not an error) if the field is not present.
+func extractFixedBytesField(data []byte, fieldNum protowire.Number, size int) ([]byte, error) {
+	pos := 0
+
+	for pos < len(data) {
+		num, wireType, tagLen := protowire.ConsumeTag(data[pos:])
+		if tagLen < 0 {
+			return nil, errors.New("malformed protobuf tag")
+		}
+		pos += tagLen
+
+		if num == fieldNum {
+			if wireType != protowire.BytesType {
+				return nil, errors.New("wrong wire type for field")
+			}
+			msgBytes, n := protowire.ConsumeBytes(data[pos:])
+			if n < 0 {
+				return nil, errors.New("invalid bytes in field")
+			}
+			if len(msgBytes) == 0 {
+				return nil, nil // proto3 zero-value encoding, treat as absent
+			}
+			if len(msgBytes) != size {
+				return nil, errors.New("field has unexpected size")
+			}
+			return msgBytes, nil
+		}
+
+		n := skipField(data[pos:], num, wireType)
+		if n < 0 {
+			return nil, errors.New("failed to skip field")
+		}
+		pos += n
+	}
+
+	return nil, nil
+}
